CID-338 refactor helper and ihubstatus
