+++ Lib/IHubStatus.js
@@ -1,0 +1,5 @@
+module.exports = {
+	IN_PROGRESS: 'IN_PROGRESS',
+	FINISHED: 'FINISHED',
+	FAILED: 'FAILED'
+};
