+++ Lib/helper.js
@@ -1,5 +1,5 @@
-module.exports = {
-	checkRegexExcludeList: function (regexExcludeList) {
+class Util {
+	static checkRegexExcludeList(regexExcludeList) {
 		if (regexExcludeList) {
 			try {
 				// checking the provided regurlar expressions for errors
@@ -10,40 +10,38 @@ module.exports = {
 			return regexExcludeList;
 		}
 		return [];
-	},
-	getISODateStringOnFromToday(daysBack = 30) {
+	}
+
+	static getISODateStringOnFromToday(daysBack = 30) {
 		const today = new Date();
 		return new Date(today.setDate(today.getDate() - daysBack)).toISOString();
-	},
-	externalId: {
-		repository: function (orgName, repoData) {
-			/**
-			 * OrgName and repo name are sanitised(no spaces) at source (GitHub)
-			 */
-			if (!orgName || !repoData) {
-				throw new Error('Failed to generate repository external ID');
-			}
-			return `${orgName}/${repoData.name}`;
-		},
-		team: function (orgName, teamData) {
-			if (!orgName || !teamData) {
-				throw new Error('Failed to generate team external ID');
-			}
-			return `${orgName}/${teamData.name}`;
-		},
-		language: function (langData) {
-			if (!langData) {
-				throw new Error('Failed to generate language external ID');
-			}
-			return `${langData.name.toLowerCase()}`;
-		}
-	},
-	iHubStatus: {
-		IN_PROGRESS: 'IN_PROGRESS',
-		FINISHED: 'FINISHED',
-		FAILED: 'FAILED'
-	},
-	iHubProgressOrigin: {
-		CONNECTOR: 'CONNECTOR'
 	}
-};
+
+	static externalId() {
+		return {
+			repository: function (orgName, repoData) {
+				/**
+				 * OrgName and repo name are sanitised(no spaces) at source (GitHub)
+				 */
+				if (!orgName || !repoData) {
+					throw new Error('Failed to generate repository external ID');
+				}
+				return `${orgName}/${repoData.name}`;
+			},
+			team: function (orgName, teamData) {
+				if (!orgName || !teamData) {
+					throw new Error('Failed to generate team external ID');
+				}
+				return `${orgName}/${teamData.name}`;
+			},
+			language: function (langData) {
+				if (!langData) {
+					throw new Error('Failed to generate language external ID');
+				}
+				return `${langData.name.toLowerCase()}`;
+			}
+		};
+	}
+}
+
+module.exports = Util;
